Add redirect for when you open login link in a new tab
Since the OmniAuth login process requires `post` requests, it doesn't work to open a login link in a new tab. Doing so bypasses the Rails javascript that uses `data-method="post"` to intercept clicks and issue them as a POST instead.

To avoid the confusing situation of a cryptic plain-text error message ("Not found. Authentication passthru."), we redirect the GET version of a login route back to the home page. This way, you can continue the login process in new tab.